LayoutTests/http/tests/security/blob-null-url-location-origin-expected.txt

CONSOLE MESSAGE: data URL frame loaded
CONSOLE MESSAGE: blob popup opened
CONSOLE MESSAGE: blob popup loadednull
CONSOLE MESSAGE: blob popup message posted
PASS
